add report and redact code
