+++ CMakeLists.txt
@@ -9,7 +9,6 @@ include_directories(include)
 
 add_subdirectory(3rdparty)
 add_subdirectory(src)
-add_subdirectory(app)
 add_subdirectory(test)
