--- app/CMakeLists.txt
@@ -1,7 +1,0 @@
-file(GLOB CPPS "*.cpp")
-file(GLOB HPPS "*.h")
-
-include_directories(../include)
-
-add_executable(app ${CPPS} ${HPPS})
-target_link_libraries(app code)
