--- app/main.cpp
@@ -1,5 +1,0 @@
-#include "PolishFromInfix.h"
-
-int main() {
- 
-}
