+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.0)
 
 file(GLOB_RECURSE CPPS "*.cpp")
-file(GLOB_RECURSE HPPS "*.h")
+file(GLOB_RECURSE HPPS "../include/*.h")
 
 include_directories(../include)
